phishing_detection_api/requirements.txt
Flask==2.1.0
Flask-Limiter==2.6.0
Flask-Caching==1.10.1
scikit-learn==0.24.2
numpy==1.21.0
joblib==1.0.1
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.0
